add AbstractComputationSteps
reduce duplication of code logic in implementations of ComputationSteps accross SQ and CE tasks in plugins
